# internal/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "coop hub",
    "version": "1"
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
    "/v1/whoami": {
      "get": {
        "summary": "Give the name and the role of the token",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "The token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "pattern": "^[a-z0-9_-]{1,64}$"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "machine",
                        "operator",
                        "orchestrator"
                      ]
                    },
                    "version": {
                      "type": "string",
                      "maxLength": 100
                    }
                  },
                  "required": [
                    "name",
                    "role",
                    "version"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "unauthorized",
                        "forbidden",
                        "not_found",
                        "conflict",
                        "ambiguous",
                        "too_large",
                        "invalid",
                        "rate_limited",
                        "unavailable"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "unauthorized",
                        "forbidden",
                        "not_found",
                        "conflict",
                        "ambiguous",
                        "too_large",
                        "invalid",
                        "rate_limited",
                        "unavailable"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "unauthorized",
                        "forbidden",
                        "not_found",
                        "conflict",
                        "ambiguous",
                        "too_large",
                        "invalid",
                        "rate_limited",
                        "unavailable"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{sid}/stream": {
      "get": {
        "summary": "Join a session and receive its events as server-sent events",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9_-]{1,64}$"
            },
            "required": true,
            "name": "sid",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^(?!(?:operator|all|any)$)[a-z0-9_-]{1,64}$"
            },
            "required": true,
            "name": "agent",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "instance",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 256
            },
            "required": true,
            "name": "host",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 4096
            },
            "required": true,
            "name": "cwd",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 64
            },
            "required": true,
            "name": "client_name",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 64
            },
            "required": true,
            "name": "client_version",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "0",
                "1"
              ]
            },
            "required": false,
            "name": "gated",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Events: joined, message, notice",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "unauthorized",
                        "forbidden",
                        "not_found",
                        "conflict",
                        "ambiguous",
                        "too_large",
                        "invalid",
                        "rate_limited",
                        "unavailable"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Not allowed: not in a session, session closed, or removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "unauthorized",
                        "forbidden",
                        "not_found",
                        "conflict",
                        "ambiguous",
                        "too_large",
                        "invalid",
                        "rate_limited",
                        "unavailable"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Unknown session or peer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "unauthorized",
                        "forbidden",
                        "not_found",
                        "conflict",
                        "ambiguous",
                        "too_large",
                        "invalid",
                        "rate_limited",
                        "unavailable"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Conflict with the session state: name taken, ambiguous peer, or yourself",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "unauthorized",
                        "forbidden",
                        "not_found",
                        "conflict",
                        "ambiguous",
                        "too_large",
                        "invalid",
                        "rate_limited",
                        "unavailable"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "unauthorized",
                        "forbidden",
                        "not_found",
                        "conflict",
                        "ambiguous",
                        "too_large",
                        "invalid",
                        "rate_limited",
                        "unavailable"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "unauthorized",
                        "forbidden",
                        "not_found",
                        "conflict",
                        "ambiguous",
                        "too_large",
                        "invalid",
                        "rate_limited",
                        "unavailable"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}